made api route for lb
